cpp02 is done ?

+++ CPP02/ex01/main.cpp
@@ -1,17 +1,32 @@
-#include <iostream>
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   main.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nsoz <nsoz@student.42.fr>                  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/08/11 18:48:03 by fgunay            #+#    #+#             */
+/*   Updated: 2024/11/13 03:33:23 by nsoz             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "Fixed.hpp"
 
 int main( void ) {
-Fixed a;
+
+Fixed       a;
 Fixed const b( 10 );
 Fixed const c( 42.42f );
 Fixed const d( b );
+
 a = Fixed( 1234.4321f );
+
 std::cout << "a is " << a << std::endl;
 std::cout << "b is " << b << std::endl;
 std::cout << "c is " << c << std::endl;
 std::cout << "d is " << d << std::endl;
-std::cout << "a is " << a.toInt() << " as integer" << std::endl;
+
+std::cout << "a is " << a.toFloat() << " as integer" << std::endl;
 std::cout << "b is " << b.toInt() << " as integer" << std::endl;
 std::cout << "c is " << c.toInt() << " as integer" << std::endl;
 std::cout << "d is " << d.toInt() << " as integer" << std::endl;
